moved get_azure_ad_users() and new_users() to provider.py
